requirements.txt
Flask==2.3.2
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
joblib==1.3.1